change IChatHub to use ResultDto and most places use Result now. IChatHubClient now is with messages instead of ids when deleting

+++ backend/Hubs/ChatHub.cs
@@ -1,9 +1,12 @@
 using System.Collections.Concurrent;
 using backend.Data.Services.Interfaces;
+using CSharpFunctionalExtensions;
 using Microsoft.AspNet.SignalR.Client.Hubs;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.SignalR;
+using Shared.Data.DTO;
 using Shared.Data.Entities;
+using Shared.Extensions;
 using Shared.Hubs.Interfaces;
 using Shared.Utils;
 
@@ -47,13 +50,17 @@ public class ChatHub : Hub<IChatHubClient>, IChatHub
         await base.OnConnectedAsync();  
     }
 
-    public override async Task OnDisconnectedAsync(Exception? exception)
+    public override async Task<Result> OnDisconnectedAsync(Exception? exception)
     {
         var claims = Context.User;
+        if (claims == null)
+        {
+            return Result.Failure("User not logged in.");
+        }
         var user = await _userManager.GetUserAsync(claims);
         if (user == null)
         {
-            return;
+            return Result.Failure("User not found.");
         }
 
         user.SignalrConnectionId = null;
@@ -61,7 +68,12 @@ public class ChatHub : Hub<IChatHubClient>, IChatHub
 
         var roomsUserIsIn = await _roomService.GetAllRoomsUserIsIn(user);
 
-        foreach (var room in roomsUserIsIn)
+        if (roomsUserIsIn.IsFailure)
+        {
+            return Result.Failure("User is not in any room");
+        }
+        
+        foreach (var room in roomsUserIsIn.Value)
         {
             await RemoveFromGroup(room.Name);
             if (room.Owner.Id == user.Id)
@@ -73,45 +85,67 @@ public class ChatHub : Hub<IChatHubClient>, IChatHub
         //await Groups.RemoveFromGroupAsync(Context.ConnectionId, claims);
 
         await base.OnDisconnectedAsync(exception);
+        return Result.Success();
     }
 
-    public async Task<bool> AddToGroup(string groupName, string password)
+    public async Task<ResultDto> AddToGroup(string groupName, string password)
     {
         password = HashUtil.Sha256String(password);
         var claims = Context.User;
         if (claims == null)
         {
-            _logger.LogError("Claims are null, user is not logged in.");
-            return false;
+            const string failure = "Claims are null, user is not logged in.";
+            _logger.LogError(failure);
+            return Result.Failure(failure).ToDto();
         }
         var user = await _userManager.GetUserAsync(claims);
         if (user == null)
         {
-            _logger.LogError("User is not found.");
-            return false;
+            const string failure = "User is not found.";
+            _logger.LogError(failure);
+            return Result.Failure(failure).ToDto();
         }
         user.SignalrConnectionId = Context.ConnectionId;
         await _userManager.UpdateAsync(user);
         
-        var room = await _roomService.Read(groupName);
-        if (room == null)
+        Room room;
+        var roomMaybe = await _roomService.Read(groupName);
+        bool isRoomNew = false;
+        if (roomMaybe.HasNoValue)
         {
-            room = await _roomService.Create(new Room
+            room = (await _roomService.Create(new Room
             {
                 Name = groupName,
                 MaxAmountOfUser = 5,
                 Owner = user,
                 Password = password,
                 Users = [user]
-            });
+            })).Value;
+            isRoomNew = true;
         }
         else
         {
+            room = roomMaybe.Value;
             if (room.Password != password)
             {
-                _logger.LogError("Passwords do not match.");
-                return false;
+                _logger.LogWarning("User '{user}' tried to join room '{room}', but passwords do not match.", user.UserName, room.Name);
+                return Result.Failure(
+                    $"User '{user.UserName}' tried to join room '{room.Name}', but passwords do not match.").ToDto();
             }
+        }
+
+        if (room.Owner.Id == user.Id && isRoomNew is false)
+        {
+            const string failure = "User is the owner of the room.";
+            _logger.LogWarning(failure);
+            return Result.Failure(failure).ToDto();
+        }
+
+        if (room.Users.Contains(user) && isRoomNew is false)
+        {
+            const string failure = "User is already in the room.";
+            _logger.LogWarning(failure);
+            return Result.Failure(failure).ToDto();
         }
 
         
@@ -123,10 +157,10 @@ public class ChatHub : Hub<IChatHubClient>, IChatHub
         };
         await CreateChatMessage(sysMessage);
         await LoadMessages(groupName);
-        return true;
+        return Result.Success().ToDto();
     }
 
-    public async Task RemoveFromGroup(string groupName)
+    public async Task<ResultDto> RemoveFromGroup(string groupName)
     {
         //TODO: delete room if last user leaves or if they are the owner.
         await Groups.RemoveFromGroupAsync(Context.ConnectionId, groupName);
@@ -135,18 +169,18 @@ public class ChatHub : Hub<IChatHubClient>, IChatHub
             Group = groupName,
             Message = $"{Context.User.Identity.Name} has left the group {groupName}."
         };
-        await CreateChatMessage(sysMessage);
+        return await CreateChatMessage(sysMessage);
     }
 
-    public async Task DeleteGroup(string groupName)
+    public async Task<ResultDto> DeleteGroup(string groupName)
     {
         var room = await _roomService.Read(groupName);
-        if (room == null)
+        if (room.HasNoValue)
         {
-            return;
+            return  Result.Failure("Room not found.").ToDto();
         }
 
-        foreach (var user in room.Users)
+        foreach (var user in room.Value.Users)
         {
             if (user.SignalrConnectionId is null)
             {
@@ -155,40 +189,56 @@ public class ChatHub : Hub<IChatHubClient>, IChatHub
             await Groups.RemoveFromGroupAsync(user.SignalrConnectionId, groupName);
         }
 
-        var messages = await _chatMessageService.GetAllGroupChatMessages(groupName);
+        var messagesResult = await _chatMessageService.GetAllGroupChatMessages(groupName);
+        if (messagesResult.IsFailure)
+        {
+            return Result.Failure(messagesResult.Error).ToDto();
+        }
+        var messages = messagesResult.Value;
         await _chatMessageService.DeleteRange(messages);
-        await _roomService.Delete(groupName);
+        var didRoomDelete = await _roomService.Delete(groupName);
+        //return Result.SuccessIf(didRoomDelete.IsSuccess, "Room was not deleted.");
+        return Result.Success().ToDto();
     }
 
-    public async Task LoadMessages(string groupName)
+    public async Task<ResultDto> LoadMessages(string groupName)
     {
-        var messages = await _chatMessageService.GetAllGroupChatMessages(groupName);
+        var messagesResult = await _chatMessageService.GetAllGroupChatMessages(groupName);
+        if (messagesResult.IsFailure)
+        {
+            return Result.Failure(messagesResult.Error).ToDto();
+        }
+        var messages = messagesResult.Value;
         messages.Sort((x, y) => y.Timestamp.CompareTo(x.Timestamp) * -1);
         
         await Clients.Group(groupName).MessagesLoaded(messages);
+        return Result.Success().ToDto();
     }
 
-    public async Task CreateChatMessage(ChatMessage chatMessage)
+    public async Task<ResultDto> CreateChatMessage(ChatMessage chatMessage)
     {
         await _chatMessageService.Create(chatMessage);
         await Clients.Group(chatMessage.Group).ChatMessageReceived(chatMessage);
+        return Result.Success().ToDto();
     }
 
-    public async Task CreateMessage(string group, string message)
+    public async Task<ResultDto> CreateMessage(string group, string message)
     {
         var claimsPrincipal = Context.User;
         
         if (claimsPrincipal == null)
         {
-            _logger.LogWarning("User is not logged in, aborting.");
-            return;
+            const string failure = "Claims are null, user is not logged in.";
+            _logger.LogWarning(failure);
+            return Result.Failure(failure).ToDto();
         }
 
         var user = await _userManager.GetUserAsync(claimsPrincipal);
         if (user == null)
         {
-            _logger.LogWarning("User is null, aborting.");
-            return;
+            const string failure = "User is not found.";
+            _logger.LogWarning(failure);
+            return Result.Failure(failure).ToDto();
         }
 
         var chatMessage = new ChatMessage
@@ -199,20 +249,36 @@ public class ChatHub : Hub<IChatHubClient>, IChatHub
         };
         await _chatMessageService.Create(chatMessage);
         await Clients.Group(chatMessage.Group).ChatMessageReceived(chatMessage);
+        return Result.Success().ToDto();
     }
 
-    public async Task ClearMessages(string groupName)
+    public async Task<ResultDto> ClearMessages(string groupName)
     {
-        var ids = (await _chatMessageService.GetAllGroupChatMessages(groupName)).Select(msg => msg.Id);
-        var guids = ids as Guid[] ?? ids.ToArray();
-        await _chatMessageService.DeleteRange(guids.ToList());
-        await Clients.Group(groupName).MessagesDeleted(guids);
+        var messagesResult = await _chatMessageService.GetAllGroupChatMessages(groupName);
+        if (messagesResult.IsFailure)
+        {
+            return Result.Failure(messagesResult.Error).ToDto();
+        }
+        var messages = messagesResult.Value;
+        await _chatMessageService.DeleteRange(messages);
+        await Clients.Group(groupName).MessagesDeleted(messages);
+        return Result.Success().ToDto();
     }
 
-    public async Task DeleteMessage(Guid id)
+    public async Task<ResultDto> DeleteMessage(Guid id)
     {
-        var chatMessage = await _chatMessageService.Read(id);
-        await _chatMessageService.Delete(id);
+        var chatMessageResult = await _chatMessageService.Read(id);
+        if (chatMessageResult.HasNoValue)
+        {
+            return Result.Failure("Couldn't find message").ToDto();
+        }
+        var chatMessage = chatMessageResult.Value;
+        var deleteResult = await _chatMessageService.Delete(id);
+        if (deleteResult.IsFailure)
+        {
+            return Result.Failure(deleteResult.Error).ToDto();
+        }
         await Clients.Group(chatMessage.Group).MessageDeleted(id);
+        return Result.Success().ToDto();
     }
 }
